requirements.txt
requests
streamlit
elasticsearch
python-dotenv